refactor(input): organize input/ into keymap/ and runtime/ subdirs
Move-only: no logic change. keymap/ holds keymaps, stationBindings,
modifiedEnterPolicy; runtime/ holds executeOutcome, managedLaunch,
paneEffects, sequenceNormalize, stationRows. The consumed hubs (router,
mouse, keyboardProtocol) and the stationInput composition root stay at
input/ root, so the ~17 external importers are untouched; only
stationOverlayLayer.ts repoints keymaps one level deeper. Relative
import depths fixed; tests moved with their source. Layout convention
documented in docs/architecture.md. Typecheck clean; 179 pass.
