+++ station/src/input/runtime/sequenceNormalize.test.ts
